stub list component
