+++ www/src/gatsby-plugin-theme-ui/components.js
@@ -12,6 +12,7 @@ import CodeBlock from "../components/code-block"
 import MdxLink from "../components/mdx-link"
 import Breakout from "../components/breakout"
 import VisuallyHidden from "../components/visually-hidden"
+import StubList from "../components/stub-list"
 
 export default {
   GuideList,
@@ -22,8 +23,9 @@ export default {
   LayerModel,
   EmailCaptureForm,
   HorizontalNavList,
-  a: MdxLink,
   Breakout,
   VisuallyHidden,
+  StubList,
+  a: MdxLink,
   pre: ({ children }) => <CodeBlock>{children}</CodeBlock>,
 }
